New folders structure for Cupydo. First step: moving files without loosing git history.
