+++ cupydoInterfaces/GetDPInterface.py
@@ -14,7 +14,7 @@
 import math
 import os
 import numpy as np
-from FSICoupler import SolidSolver
+from cupydo.FSICoupler import SolidSolver
 from getdp import *
 
 # ----------------------------------------------------------------------
